[CDCSDK] Handling tserver crash for accessing uninitialised log object pointer for SetCDCCheckpoint API.
Summary:
During the call to SetCDCCheckpoint, with the bootstrap request enabled we observed that the tsever crashed sometime. we found that during the bootstrap request CDC read the latest replicated op_id from WAL, if the log object is not initialised during that time, dereferencing an invalid pointer will cause a crash for the tsever.

As part of the fix we valid the log object initialisation before we dereference it.

Test Plan: Running all the c and java testcases

Reviewers: skumar, abharadwaj, aagrawal, vkushwaha

Subscribers: ycdcxcluster, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D18055