[#25455] YSQL: fix style bad opening parenthesis
Summary:
Add a new pg lint rule bad_opening_paren to catch cases where lines end
on an opening parenthesis.  This is commonly the case for function
calls.  YB folks often break on the opening parenthesis and put all
arguments on one indent level ahead of the previous line.  This is not
seen in upstream postgres code, and pgindent would align all the
arguments to the opening parenthesis, making it look awkward (tabs are
expanded in the below example):

    somefunction(
                 firstarg,
                 secondarg);

Fix all failures of this lint rule manually.  This brings us a step
closer in being able to reasonably use pgindent.
Jira: DB-14700

Test Plan:
On Almalinux 8:

    [ -z "$(arc lint src/postgres/**/*.{c,h} | grep bad_opening_paren)" ]

Close: #25455

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: yql, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D40933